Moved display to new module with OO structure
Code is MUCH cleaner now :)
